+++ lib/presentation/widgets/compact_file_preview.dart
