+++ lib/presentation/widgets/message_bubble.dart
