+++ lib/presentation/widgets/transcription_bubble.dart
