+++ lib/rtc/room.dart
@@ -11,8 +11,8 @@ import 'package:daakia_vc_flutter_sdk/rtc/widgets/participant.dart';
 import 'package:daakia_vc_flutter_sdk/rtc/widgets/participant_info.dart';
 import 'package:daakia_vc_flutter_sdk/rtc/widgets/pip_screen.dart';
 import 'package:daakia_vc_flutter_sdk/rtc/widgets/rtc_controls.dart';
-import 'package:daakia_vc_flutter_sdk/screens/customWidget/emoji_reaction_widget.dart';
-import 'package:daakia_vc_flutter_sdk/utils/exts.dart';
+import 'package:daakia_vc_flutter_sdk/presentation/widgets/emoji_reaction_widget.dart';
+import 'package:daakia_vc_flutter_sdk/utils/rtc_ext.dart';
 import 'package:daakia_vc_flutter_sdk/utils/storage_helper.dart';
 import 'package:daakia_vc_flutter_sdk/viewmodel/rtc_provider.dart';
 import 'package:daakia_vc_flutter_sdk/viewmodel/rtc_viewmodel.dart';
@@ -25,7 +25,7 @@ import 'package:wakelock_plus/wakelock_plus.dart';
 
 import '../model/emoji_message.dart';
 import '../model/remote_activity_data.dart';
-import '../screens/bottomsheet/transcription_screen.dart';
+import '../presentation/pages/transcription_screen.dart';
 import '../utils/constants.dart';
 import '../utils/meeting_actions.dart';
 import '../utils/utils.dart';
