+++ lib/utils/rtc_ext.dart
@@ -1,0 +1,91 @@
+import 'package:flutter/material.dart';
+
+extension RTCExampleExt on BuildContext {
+  //
+  Future<bool?> showPublishDialog() => showDialog<bool>(
+        context: this,
+        builder: (ctx) => AlertDialog(
+          title: const Text('Publish'),
+          content: const Text('Would you like to publish your Camera & Mic ?'),
+          actions: [
+            TextButton(
+              onPressed: () => Navigator.pop(ctx, false),
+              child: const Text('NO'),
+            ),
+            TextButton(
+              onPressed: () => Navigator.pop(ctx, true),
+              child: const Text('YES'),
+            ),
+          ],
+        ),
+      );
+
+  Future<bool?> showPlayAudioManuallyDialog() => showDialog<bool>(
+        context: this,
+        builder: (ctx) => AlertDialog(
+          title: const Text('Play Audio'),
+          content: const Text(
+              'You need to manually activate audio PlayBack for iOS Safari !'),
+          actions: [
+            TextButton(
+              onPressed: () => Navigator.pop(ctx, false),
+              child: const Text('Ignore'),
+            ),
+            TextButton(
+              onPressed: () => Navigator.pop(ctx, true),
+              child: const Text('Play Audio'),
+            ),
+          ],
+        ),
+      );
+
+  Future<void> showErrorDialog(dynamic exception) => showDialog<void>(
+        context: this,
+        builder: (ctx) => AlertDialog(
+          title: const Text('Error'),
+          content: Text(exception.toString()),
+          actions: [
+            TextButton(
+              onPressed: () => Navigator.pop(ctx),
+              child: const Text('OK'),
+            )
+          ],
+        ),
+      );
+
+  Future<bool?> showDisconnectDialog() => showDialog<bool>(
+        context: this,
+        builder: (ctx) => AlertDialog(
+          title: const Text('Disconnect'),
+          content: const Text('Are you sure to disconnect?'),
+          actions: [
+            TextButton(
+              onPressed: () => Navigator.pop(ctx, false),
+              child: const Text('Cancel'),
+            ),
+            TextButton(
+              onPressed: () => Navigator.pop(ctx, true),
+              child: const Text('Disconnect'),
+            ),
+          ],
+        ),
+      );
+
+  Future<bool?> showPermissionAskDialog(String message) => showDialog<bool>(
+    context: this,
+    builder: (ctx) => AlertDialog(
+      backgroundColor: Colors.grey[900],
+      content: Text(message, style: const TextStyle(color: Colors.white),),
+      actions: [
+        TextButton(
+          onPressed: () => Navigator.pop(ctx, false),
+          child: const Text('Reject', style: TextStyle(color: Colors.white),),
+        ),
+        TextButton(
+          onPressed: () => Navigator.pop(ctx, true),
+          child: const Text('Accept', style: TextStyle(color: Colors.white),),
+        ),
+      ],
+    ),
+  );
+}
