+++ lib/viewmodel/rtc_viewmodel.dart
@@ -8,7 +8,7 @@ import 'package:daakia_vc_flutter_sdk/events/rtc_events.dart';
 import 'package:daakia_vc_flutter_sdk/model/remote_activity_data.dart';
 import 'package:daakia_vc_flutter_sdk/model/transcription_action_model.dart';
 import 'package:daakia_vc_flutter_sdk/model/transcription_model.dart';
-import 'package:daakia_vc_flutter_sdk/utils/language_json.dart';
+import 'package:daakia_vc_flutter_sdk/resources/json/language_json.dart';
 import 'package:daakia_vc_flutter_sdk/utils/utils.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
